Adding a webpack based build process
