+++ package.json
@@ -1,0 +1,15 @@
+{
+  "name": "unanet-summarizer",
+  "version": "0.1.0",
+  "scripts": {
+    "start": "webpack-dev-server --open --config webpack.config.development.js",
+    "build": "webpack"
+  },
+  "devDependencies": {
+    "copy-webpack-plugin": "^5.0.2",
+    "raw-loader": "^2.0.0",
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.1",
+    "webpack-dev-server": "^3.3.1"
+  }
+}
