+++ unanet-summarizer.js
@@ -1,230 +1,1 @@
-window.summarizeUnanetTime = (function() {
-    var Template = function(html) {
-        // see: http://krasimirtsonev.com/blog/article/Javascript-template-engine-in-just-20-line
-        var re = /<%([^%>]+)?%>/g,
-            reExp = /(^( )?(if|for|else|switch|case|break|{|}))(.*)?/g,
-            code = 'var r=[];\n',
-            cursor = 0,
-            match;
-
-        var add = function(line, js) {
-            js ? (code += line.match(reExp) ? line + '\n' : 'r.push(' + line + ');\n')
-               : (code += line != '' ? 'r.push("' + line.replace(/"/g, '\\"') + '");\n' : '');
-            return add;
-        };
-
-        while(match = re.exec(html)) {
-            add(html.slice(cursor, match.index))(match[1], true);
-            cursor = match.index + match[0].length;
-        }
-
-        add(html.substr(cursor, html.length - cursor));
-        code += 'return r.join("");';
-
-        return new Function(code.replace(/[\r\t\n]/g, ''));
-    };
-
-    const ASSUMED_HOURS_PER_DAY = 8;
-    const SUMMARIZER_ROOT = 'https://excellalabs.github.io/unanet-summarizer/';
-    const SUMMARIZER_STYLESHEET = SUMMARIZER_ROOT + '/summarizer-style.css';
-
-    const CONTAINER_ID = 'unanet-summarizer';
-    const STYLESHEET_ID = 'unanet-summarizer-style';
-    const CSS_CLASS = 'unanet-summary';
-
-    const CONTAINER_TEMPLATE = Template(
-      '<h2>Unanet Time Summary</h2>' +
-      '<table>' +
-        '<thead>' +
-          '<tr>' +
-            '<th colspan="<% this.hoursByProjectType.length %>">Project Types</th>' +
-            '<th colspan="3">Totals</th>' +
-            '<th colspan="3">Pay Period Tracking</th>' +
-          '</tr>' +
-          '<tr>' +
-            '<% for (var i in this.hoursByProjectType) { %>' +
-              '<th><% this.hoursByProjectType[i].projectType %></th>' +
-            '<% } %>' +
-            '<th>+ Hours</th>' +
-            '<th>Non + Hours</th>' +
-            '<th>Grand Total</th>' +
-            '<th>Potential Hours</th>' +
-            '<th>Your + Hours</th>' +
-            '<th>Tracking</th>' +
-          '</tr>' +
-        '</thead>' +
-        '<tbody>' +
-          '<tr>' +
-            '<% for (var i in this.hoursByProjectType) { %>' +
-              '<td><% this.hoursByProjectType[i].totalHours %></td>' +
-            '<% } %>' +
-            '<td><% this.totalPlusHoursResult %></td>' +
-            '<td><% this.totalNonPlusHoursResult %></td>' +
-            '<td><% this.totalHoursResult %></td>' +
-            '<td><% this.hoursTargetForPayPeriod %></td>' +
-            '<td><% this.totalPlusHoursResult %></td>' +
-            '<td><% this.hoursForTracking %></td>' +
-          '</tr>' +
-        '</tbody>' +
-      '</table>'
-    );
-
-    var isReadOnlyTimesheet = function() {
-        // if there are no inputs, the timesheet is readonly
-        return document.querySelectorAll('input').length === 0;
-    };
-
-    var toArray = function(nodeList) {
-        return [].slice.call(nodeList);        
-    };
-
-    const IsReadOnly = isReadOnlyTimesheet();
-
-
-    var obtainTimeEntryRows = function() { 
-        var arrayToReturn = []; 
-
-        var rows = toArray(IsReadOnly ?
-          document.querySelectorAll("table.timesheet > tbody:first-of-type > tr")
-          : document.querySelectorAll("#timesheet > tbody:first-of-type > tr")
-        );
-
-        return rows
-          .map(function(timesheetRow) {
-            var projectType;
-            var timeValue;
-        
-            if (IsReadOnly) {
-                projectType = timesheetRow.querySelector(':nth-child(4)').textContent || "";
-                timeValue = parseFloat(timesheetRow.querySelector(':last-child').textContent) || parseFloat(0.0);
-            } else {
-                projectType = timesheetRow.querySelector("td.project-type > select > option:checked").text;  
-                timeValue = parseFloat(timesheetRow.querySelector('td.total > input').getAttribute('value')) || parseFloat(0.0); 
-            }
-        
-            return (!projectType || projectType === '') ? null : { projectType: projectType, timeValue: timeValue };
-          })
-          .filter(function(timesheetRow) {
-            return timesheetRow !== null;
-          });
-    };
-
-    var totalHoursReduceFunction = function(acc, obj) {
-        return acc + obj.timeValue;
-    };
-
-    var totalHoursByProjectType = function(acc, obj){ 
-        if(!acc.find(function(element){return element.projectType === obj.projectType;})){
-            acc.push({projectType: obj.projectType, totalHours: 0.0});
-        }
-        
-        var entry = acc.find(function(element){return element.projectType === obj.projectType;});
-        entry.totalHours += obj.timeValue;
-        
-        return acc;
-    };
-
-    var totalPlusHours = function(acc, obj){
-        if (obj.projectType.includes("+")){
-            acc = acc + obj.timeValue;
-        }
-        return acc;
-    };
-
-    var totalNonPlusHours = function(acc, obj){
-        if (!obj.projectType.includes("+")){
-            acc = acc + obj.timeValue;
-        }
-        return acc;
-    };
-
-    var createContainer = function() {
-        var container = document.createElement('div');
-        container.id = CONTAINER_ID;
-        container.className = CSS_CLASS;
-        return document.body.insertBefore(container, document.body.firstChild);
-    };
-
-    var createStylesheetRef = function() {
-        var link = document.createElement('link');
-        link.id = STYLESHEET_ID;
-        link.rel = 'stylesheet';
-        link.type = 'text/css';
-        link.href = SUMMARIZER_STYLESHEET;
-        return document.head.appendChild(link);
-    };
-
-    // This function returns an object with a reducer function (a way of reducing an array of hours which we'll pass later)
-    // and the starting state of the accumulators that those reducers will use. 
-    var getReducers = function() {
-        return {
-            hoursByProjectType: { fn: totalHoursByProjectType, init: [] },
-            totalPlusHoursResult: { fn: totalPlusHours, init: 0.0 },
-            totalNonPlusHoursResult: { fn: totalNonPlusHours, init: 0.0 },
-            totalHoursResult: { fn: totalHoursReduceFunction, init: 0.0 }
-        };
-    };
-
-
-    var getBusinessDatesCount = function(startDate, endDate) {
-        // provided from https://stackoverflow.com/a/37069277/316847
-        var count = 0;
-        var curDate = startDate;
-        while (curDate <= endDate) {
-            var dayOfWeek = curDate.getDay();
-            if(!((dayOfWeek == 6) || (dayOfWeek == 0)))
-               count++;
-            curDate.setDate(curDate.getDate() + 1);
-        }
-        return count;
-    };
-
-    var getWeekdaysInTimesheet = function(){
-        if (IsReadOnly) {
-            return document.querySelectorAll('table.timesheet > tbody > tr:first-of-type > td.weekday').length;
-        }
-        else {
-            return document.querySelectorAll('#timesheet > tbody > tr:first-of-type > td.weekday-hours').length;
-        }
-    };
-
-    var getDaysLeftInTimesheet = function() {
-        if (IsReadOnly){ return 0; }
-        else {
-            var dateElements = document.querySelectorAll('span.dom');
-            var lastDateOnTimesheet = parseInt(dateElements[dateElements.length - 1].textContent);
-        
-            var today = new Date();
-            var fullLastDate = new Date(today.getFullYear(), today.getMonth(), lastDateOnTimesheet);
-            var fullTodayDate = new Date(today.getFullYear(), today.getMonth(), today.getDate());
-
-            return getBusinessDatesCount(fullTodayDate, fullLastDate);
-        }
-    }
-
-    // Execution
-    return function() {
-        // inject stylesheet
-        if (!document.getElementById(STYLESHEET_ID)) {
-            createStylesheetRef();
-        }
-
-        var timeEntries = obtainTimeEntryRows();        
-        var reducers = getReducers();
-
-        // Takes each reducer function name and function, calls reduce using that function,
-        // and sets the result on an object that we reference in our document generation.
-        var properties = Object.keys(reducers).reduce(function(acc, property) {
-            var config = reducers[property];
-            acc[property] = timeEntries.reduce(config.fn, config.init);
-            return acc;
-        }, {});
-
-        properties.hoursTargetForPayPeriod = getWeekdaysInTimesheet() * ASSUMED_HOURS_PER_DAY;
-        properties.hoursForTracking = -(properties.hoursTargetForPayPeriod - (getDaysLeftInTimesheet() * ASSUMED_HOURS_PER_DAY) - properties.totalPlusHoursResult);
-
-       
-        var container = document.getElementById(CONTAINER_ID) || createContainer();
-        container.innerHTML = CONTAINER_TEMPLATE.apply(properties);
-    };
-})();
+!function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=0)}([function(t,e,n){const r=n(1),o=n(2).default,u=n(3).default,a=n(4);window.unanetSummarizer=window.unanetSummarizer||{},window.unanetSummarizer.generateSummary=function(){const t="unanet-summarizer",e="unanet-summarizer-style",n="unanet-summary",i=r(o);return function(){var r;document.getElementById(e)||((r=document.createElement("style")).id=e,r.appendChild(document.createTextNode(u)),document.head.appendChild(r)),(document.getElementById(t)||function(){var e=document.createElement("div");return e.id=t,e.className=n,document.body.insertBefore(e,document.body.firstChild)}()).innerHTML=i.apply(a())}}(),window.summarizeUnanetTime=window.unanetSummarizer.generateSummary},function(t,e){t.exports=function(t){for(var e,n=/<%([^%>]+)?%>/g,r=/(^( )?(if|for|else|switch|case|break|{|}))(.*)?/g,o="var r=[];\n",u=0,a=function(t,e){return o+=e?t.match(r)?t+"\n":"r.push("+t+");\n":""!=t?'r.push("'+t.replace(/"/g,'\\"')+'");\n':"",a};e=n.exec(t);)a(t.slice(u,e.index))(e[1],!0),u=e.index+e[0].length;return a(t.substr(u,t.length-u)),o+='return r.join("");',new Function(o.replace(/[\r\t\n]/g,""))}},function(t,e,n){"use strict";n.r(e),e.default='<h2>Unanet Time Summary</h2>\n<table>\n  <thead>\n    <tr>\n      <th colspan="<% this.hoursByProjectType.length %>">Project Types</th>\n      <th colspan="3">Totals</th>\n      <th colspan="3">Pay Period Tracking</th>\n    </tr>\n    <tr>\n      <% for (var i in this.hoursByProjectType) { %>\n        <th><% this.hoursByProjectType[i].projectType %></th>\n      <% } %>\n      <th>+ Hours</th>\n      <th>Non + Hours</th>\n      <th>Grand Total</th>\n      <th>Potential Hours</th>\n      <th>Your + Hours</th>\n      <th>Tracking</th>\n    </tr>\n  </thead>\n  <tbody>\n    <tr>\n      <% for (var i in this.hoursByProjectType) { %>\n        <td><% this.hoursByProjectType[i].totalHours %></td>\n      <% } %>\n      <td><% this.totalPlusHoursResult %></td>\n      <td><% this.totalNonPlusHoursResult %></td>\n      <td><% this.totalHoursResult %></td>\n      <td><% this.hoursTargetForPayPeriod %></td>\n      <td><% this.totalPlusHoursResult %></td>\n      <td><% this.hoursForTracking %></td>\n    </tr>\n  </tbody>\n</table>\n'},function(t,e,n){"use strict";n.r(e),e.default=".unanet-summary {\n  padding: 0 0 20px 20px;\n}\n\n.unanet-summary table {\n  border: 2px solid;\n}\n"},function(t,e){t.exports=function(){const t=0===document.querySelectorAll("input").length;var e=function(t,e){return t+e.timeValue},n=function(t,e){return t.find(function(t){return t.projectType===e.projectType})||t.push({projectType:e.projectType,totalHours:0}),t.find(function(t){return t.projectType===e.projectType}).totalHours+=e.timeValue,t},r=function(t,e){return e.projectType.includes("+")&&(t+=e.timeValue),t},o=function(t,e){return e.projectType.includes("+")||(t+=e.timeValue),t};return function(){var u,a=(u=t?document.querySelectorAll("table.timesheet > tbody:first-of-type > tr"):document.querySelectorAll("#timesheet > tbody:first-of-type > tr"),[].slice.call(u)).map(function(e){var n,r;return t?(n=e.querySelector(":nth-child(4)").textContent||"",r=parseFloat(e.querySelector(":last-child").textContent)||parseFloat(0)):(n=e.querySelector("td.project-type > select > option:checked").text,r=parseFloat(e.querySelector("td.total > input").getAttribute("value"))||parseFloat(0)),n&&""!==n?{projectType:n,timeValue:r}:null}).filter(function(t){return null!==t}),i={hoursByProjectType:{fn:n,init:[]},totalPlusHoursResult:{fn:r,init:0},totalNonPlusHoursResult:{fn:o,init:0},totalHoursResult:{fn:e,init:0}},l=Object.keys(i).reduce(function(t,e){var n=i[e];return t[e]=a.reduce(n.fn,n.init),t},{});return l.hoursTargetForPayPeriod=8*(t?document.querySelectorAll("table.timesheet > tbody > tr:first-of-type > td.weekday").length:document.querySelectorAll("#timesheet > tbody > tr:first-of-type > td.weekday-hours").length),l.hoursForTracking=-(l.hoursTargetForPayPeriod-8*function(){if(t)return 0;var e=document.querySelectorAll("span.dom"),n=parseInt(e[e.length-1].textContent),r=new Date,o=new Date(r.getFullYear(),r.getMonth(),n);return function(t,e){for(var n=0,r=t;r<=e;){var o=r.getDay();6!=o&&0!=o&&n++,r.setDate(r.getDate()+1)}return n}(new Date(r.getFullYear(),r.getMonth(),r.getDate()),o)}()-l.totalPlusHoursResult),l}}()}]);
